[aws-datastore] Re-enable test for cloud to local synchronization
blogOwnerCreatedAndUpdatedRemotelyIsFoundLocally had been @Ignored,
until now. In order to arrange a creation and an update on the AppSync
API, it is necessary to pass a version number in the update call. That
had been possible until recently.

Additionally, the receipt of a cloud event must be announced over Hub.
This way, the test has something for which to listen.