[aws-datastore] Mutation outbox precedence over cloud content
While applying cloud content back into the local storage, the Merger
must consider whether or not the Mutation Outbox has a change record
corresponding to a model with the same UUID.

If the Mutation Outbox has a pending change for a model of the same ID,
then the Merger shall NOT apply the cloud update. This is to prevent
loss of customer data on the client.