[aws-datastore] Single responsibility for SynchronousStorageAdapter
Littered around the test code were various sychronous operations on top
of a LocalStorageAdapter. Likewise, the InMemoryStorageAdapter had grown
a couple of synchronous operations which it exposed in tests. The
responsibility for syncrhonous access to a storage adapter was split
into a few different places, *including* the utility that is meant to
achieve this, the SynchronousStorageAdapter.

To resolve this, numerous tests are refactored to make singular
reference to the SynchronousStorageAdapter only, and in liue of the
other previously mechanisms of doing syncrhonous storage access.