[aws-datastore] AWSDataStorePlugin via Constructor (#397)
Disposition 1. All of the other plugins currently used in Amplify can be
created by constructor. The `AWSDataStorePlugin.forModels(...)` factory
breaks the customer's expectations of how plugins are instantiated.

Disposition 2. It is possible to dynamically locate the code-generated
model provider, internal to the plugin, and use it to instantiate the
plugin. To simplify the happy path of on-boarding with the DataStore, a
no-arguments constructor is provided.

The plugin may now be added into the system by the same means as the
other plugins:

    Amplify.addPlugin(new AWSDataStorePlugin());
    Amplify.configure(getApplicationContext());