[aws-api] Reinstate List, Separate Date and AWSTimestamp
Recent changes to request/response handling broke the integration tests.

Firstly, we must handle List repsonses. The List.class type adapter is
added back to the Gson instance that deserializes respones.

Secondly, the API integration tests rely upon the existence of a Date
type, with some existing assumptions about how this type is used.

Until the code-gen can be updated, we need to maintain the same
assumptions around the use of Date. At that time, we could deprecate it
as a type.

To faciliate this, a new AWSTimestamp model is created, which is a type
that has not yet been used in this project.

The contract on the Date type is maintained, so that our integraiton
tests will continue to pass.