[aws-datastore] PendingMutations are now Comparable according to their creation times (#439)
Changes the ID type of the `PendingMutation`. The `PendingMutation` now
has its own UUID. A mechanism still exists to obtain the ID of the
associated model:

    pendingMutation.getMutationId();
    pendingMutation.getMutatedItem().getId();

The `PendingMuation.PersistentRecord` now has *two* IDs, a "mutation
ID", and a "decoded model ID":

    record.getId();
    record.getDecodedModelId();

Previously, the record's `getId()` would return the same ID as
`pendingMutation.getMutatedItem().getId()`.

After this change, the mutation ID is unique to the PendingMutation
itself. This ID is also of a different type. It is a time-based UUIDv1.
This type of UUID was chosen as it may be sorted according to its
timestamp. And as a consequence, `PendingMutation`s themselves may now
be sorted by their UUIDv1s. `PendingMutation.PersistentRecord`s use this
as their primary `getId()` value.

The old ID of `PendingMutation.PersistentRecord` is now available by
`getDecodedModelIs()`. This is the ID of the _associated model_ -- the
thing that had undergone some mutation, to begin with. As before, this
UUID is still a v4 UUID, maintained for its technical superiority over
the time-based V1. This value is stored on the record, so that we can
still query for records based on _model ID_, which is what is generally
used around the DataStore system.