Simplify Hub and HubAccumulator (#448)
Greatly simplifies the logic in the `AWSHubPlugin` and `HubAccumulator`
utility, towards the goal of reducing multi-threading race conditions.

The `HubInstrumentedTest` is transitioned to `AWSHubPluginTest`, a unit
test. `HubAccumulatorTest` is added.

Tests are updated to use the simplified `HubAccumulator` pattern, which no
longer accepts a dynamic quantity of events to await.