Fix IDE integrations for Windows
The verifyNode task only worked on macOS and Linux, and really isn't
necessary anyway. Kill it, and instead just try to run the command. If
it's missing, throw an error.

Also, throw an exception instead of printing the error. This has two
advantages: 1) the error is red and angry and obvious 2) the user can
try again via Android Studio without modifying the gradle.build file